Bookcases editions renamed to viewer
