+++ app/src/main/res/layout/bookcase_editions_layout.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".ui.modules.bookcases.editions.BookcaseEditionsActivity"
+    tools:context=".ui.modules.bookcases.viewer.BookcaseViewerActivity"
     >
 
     <android.support.design.widget.CoordinatorLayout
