+++ app/src/main/res/layout/bookcase_editor_layout.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".ui.modules.bookcases.editions.BookcaseEditionsActivity"
+    tools:context=".ui.modules.bookcases.editor.BookcaseEditorActivty"
     >
 
     <LinearLayout
